Added verify account section to edit profile page and refactored
