+++ components/shared/index.tsx
@@ -1,4 +1,6 @@
 import { LabeledIcon } from "./LabeledIcon"
+import { PhoneVerificationModal } from "./PhoneVerificationModal"
 import TitledSectionCard from "./TitledSectionCard"
+import { VerifyAccountSection } from "./VerifyAccountSection"
 
-export { LabeledIcon, TitledSectionCard }
+export { LabeledIcon, PhoneVerificationModal, TitledSectionCard, VerifyAccountSection }
