+++ public/locales/en/auth.json
@@ -59,5 +59,38 @@
   "pleaseConsider": "Please consider logging in first:",
   "signInToAccess": "Login required to access the content",
   "almostThere": "Almost there!",
-  "justLogIn": "You’ll be redirected automatically after logging in."
+  "justLogIn": "You’ll be redirected automatically after logging in.",
+  "verifyAccountSection": {
+    "title": "Verify Account",
+    "description": "MAPLE requires our users to verify both their <bold>email address</bold> and <bold>phone number</bold> to be able to post testimony on our platform. We do this in an effort to reduce the number of bots that are able to publish on MAPLE.",
+    "emailHeader": "Email address",
+    "verifyEmail": "Click the link in your email inbox from MAPLE.",
+    "verifyEmailSpam": "If you don’t see it, be sure to check your spam folder.",
+    "checkEmail": "Check your email!",
+    "sendAnotherLink": "Resend link",
+    "phoneHeader": "Phone number",
+    "phoneDescription": "Receive a verification code on your phone",
+    "verifyPhone": "Verify",
+    "successMessage": "Successfully verified. Thank you!"
+  },
+  "phoneVerificationModalTitle": "Verify Phone",
+  "phoneVerification": {
+    "stepOne": "Step one.",
+    "stepOneDescription": "Enter your phone number below.",
+    "phonePlaceholder": "Enter your 10-digit phone number",
+    "receiveCode": "Receive code",
+    "stepTwo": "Step two.",
+    "stepTwoDescription": "Enter the code that was sent to your phone.",
+    "codePlaceholder": "Enter six-digit code",
+    "verify": "Verify",
+    "errors": {
+      "credentialAlreadyInUse": "This phone number is already linked to another account.",
+      "providerAlreadyLinked": "This account already has a phone number linked.",
+      "invalidPhoneNumber": "Please enter a valid phone number\n(e.g. 617 555-1234).",
+      "operationNotAllowed": "Phone verification is not enabled. Please try again later or contact us at info@mapletestimony.org."
+    },
+    "signedInRequired": "You must be signed in to verify your phone.",
+    "enterVerificationCode": "Please enter the verification code.",
+    "successMessage": "Success! Your phone number has been verified."
+  }
 }
